pyproject.toml
[project]
name = "unified-agentic-concierge"
version = "0.1.0"
description = "Multi-Agent Hospitality Intelligence Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.30",
    "langchain-google-genai>=1.0.0",
    "neo4j>=5.19.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "confluent-kafka>=2.3.0",
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "python-dotenv>=1.0.1",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.6",
    "black>=24.3.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
